fix migration foreign key types

+++ database/migrations/2025_09_09_000002_create_asset_images_table.php
@@ -9,7 +9,7 @@ return new class extends Migration {
     {
         Schema::create('asset_images', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('asset_id');
+            $table->unsignedInteger('asset_id');
             $table->string('file_path');
             $table->string('caption')->nullable();
             $table->timestamps();
